Fix flowr package typo
