+++ src/main/java/org/rumbledb/compiler/RuntimeIteratorVisitor.java
@@ -89,15 +89,15 @@ import org.rumbledb.runtime.control.IfRuntimeIterator;
 import org.rumbledb.runtime.control.SwitchRuntimeIterator;
 import org.rumbledb.runtime.control.TypeswitchRuntimeIterator;
 import org.rumbledb.runtime.control.TypeswitchRuntimeIteratorCase;
-import org.rumbledb.runtime.flowr.CountClauseSparkIterator;
-import org.rumbledb.runtime.flowr.ForClauseSparkIterator;
-import org.rumbledb.runtime.flowr.GroupByClauseSparkIterator;
-import org.rumbledb.runtime.flowr.LetClauseSparkIterator;
-import org.rumbledb.runtime.flowr.OrderByClauseSparkIterator;
-import org.rumbledb.runtime.flowr.ReturnClauseSparkIterator;
-import org.rumbledb.runtime.flowr.WhereClauseSparkIterator;
-import org.rumbledb.runtime.flowr.expression.GroupByClauseSparkIteratorExpression;
-import org.rumbledb.runtime.flowr.expression.OrderByClauseAnnotatedChildIterator;
+import org.rumbledb.runtime.flwor.clauses.CountClauseSparkIterator;
+import org.rumbledb.runtime.flwor.clauses.ForClauseSparkIterator;
+import org.rumbledb.runtime.flwor.clauses.GroupByClauseSparkIterator;
+import org.rumbledb.runtime.flwor.clauses.LetClauseSparkIterator;
+import org.rumbledb.runtime.flwor.clauses.OrderByClauseSparkIterator;
+import org.rumbledb.runtime.flwor.clauses.ReturnClauseSparkIterator;
+import org.rumbledb.runtime.flwor.clauses.WhereClauseSparkIterator;
+import org.rumbledb.runtime.flwor.expression.GroupByClauseSparkIteratorExpression;
+import org.rumbledb.runtime.flwor.expression.OrderByClauseAnnotatedChildIterator;
 import org.rumbledb.runtime.functions.DynamicFunctionCallIterator;
 import org.rumbledb.runtime.functions.FunctionRuntimeIterator;
 import org.rumbledb.runtime.functions.StaticUserDefinedFunctionCallIterator;
