+++ src/main/java/org/rumbledb/runtime/flwor/expression/GroupByClauseSparkIteratorExpression.java
@@ -18,48 +18,42 @@
  *
  */
 
-package org.rumbledb.runtime.flowr.expression;
+package org.rumbledb.runtime.flwor.expression;
 
-import org.rumbledb.expressions.flowr.OrderByClauseExpr;
+
+import org.rumbledb.exceptions.ExceptionMetadata;
 import org.rumbledb.runtime.RuntimeIterator;
+import org.rumbledb.runtime.primary.VariableReferenceIterator;
 
 import java.io.Serializable;
 
-public class OrderByClauseAnnotatedChildIterator implements Serializable {
+public class GroupByClauseSparkIteratorExpression implements Serializable {
+
 
     private static final long serialVersionUID = 1L;
-    private final RuntimeIterator iterator;
-    private final boolean ascending;
-    private final String uri;
-    private final OrderByClauseExpr.EMPTY_ORDER emptyOrder;
+    private final VariableReferenceIterator variableReference;
+    private final RuntimeIterator expression;
+    private final ExceptionMetadata iteratorMetadata;
 
-
-    public OrderByClauseAnnotatedChildIterator(
-            RuntimeIterator iterator,
-            boolean ascending,
-            String uri,
-            OrderByClauseExpr.EMPTY_ORDER empty_order
+    public GroupByClauseSparkIteratorExpression(
+            RuntimeIterator expression,
+            VariableReferenceIterator variable,
+            ExceptionMetadata iteratorMetadata
     ) {
-        this.iterator = iterator;
-        this.ascending = ascending;
-        this.uri = uri;
-        this.emptyOrder = empty_order;
+        this.expression = expression;
+        this.variableReference = variable;
+        this.iteratorMetadata = iteratorMetadata;
     }
 
-    public RuntimeIterator getIterator() {
-        return this.iterator;
+    public VariableReferenceIterator getVariableReference() {
+        return this.variableReference;
     }
 
-    public boolean isAscending() {
-        return this.ascending;
+    public ExceptionMetadata getIteratorMetadata() {
+        return this.iteratorMetadata;
     }
 
-    public String getUri() {
-        return this.uri;
+    public RuntimeIterator getExpression() {
+        return this.expression;
     }
-
-    public OrderByClauseExpr.EMPTY_ORDER getEmptyOrder() {
-        return this.emptyOrder;
-    }
-
 }
